fix(indochain): convert multi-height test height type

+++ IndoChain/internal/p2p/consensus_runtime_multinode_test.go
@@ -388,7 +388,7 @@ func TestInMemoryTransportRuntimeFinalizedBlockMultiHeight(t *testing.T) {
 
 	var previousHash = n.HeadHash
 	for height := types.Height(1); height <= 2; height++ {
-		state, err := consensus.NewRoundState(devnet.ProtocolVersion, devnet.ChainID, height, 0)
+		state, err := consensus.NewRoundState(devnet.ProtocolVersion, devnet.ChainID, uint64(height), 0)
 		if err != nil {
 			t.Fatal(err)
 		}
